feat: 리뷰 최대 점수 제한

+++ Movie/src/main/kotlin/com/example/Movie/review/dto/createMovieResponse.kt
@@ -2,6 +2,6 @@ package com.example.Movie.review.dto
 
 data class createMovieResponse (
     val content: String,
-    val grade: Int,
+    val grade: Float,
     val movie: String
     )
